refactor: guard -> auth 이름 바꿈
