+++ backend/src/common/auth/ft-auth/index.ts
