+++ backend/src/common/auth/jwt-auth/index.ts
@@ -1,0 +1,4 @@
+export * from './jwt-auth.guard';
+export * from './jwtPayload';
+export * from './req-user';
+export * from './socket-user';
