+++ backend/src/common/auth/jwt-auth/jwtPayload.ts
@@ -1,3 +1,4 @@
 export type JwtPayload = {
   id: number;
+  intraId: string;
 };
