+++ backend/src/common/auth/jwt-auth/req-user.ts
@@ -1,10 +1,10 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { JwtPayload } from './jwtPayload';
+import { User } from '../../database/entities/user.entity';
 
 /**
  * @description Request에 담긴 User를 가져온다.
  */
-export const ReqJwtPayload = createParamDecorator((data, ctx: ExecutionContext): JwtPayload => {
+export const ReqUser = createParamDecorator((data, ctx: ExecutionContext): User => {
   const req = ctx.switchToHttp().getRequest();
 
   return req.user;
