+++ backend/src/common/auth/jwt-auth/socket-user.ts
@@ -1,10 +1,10 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { JwtPayload } from './jwtPayload';
+import { User } from '../../database/entities/user.entity';
 
 /**
  * @description Client에 담긴 User를 가져온다.
  */
-export const SocketJwtPayload = createParamDecorator((data, ctx: ExecutionContext): JwtPayload => {
+export const SocketUser = createParamDecorator((data, ctx: ExecutionContext): User => {
   const req = ctx.switchToWs().getClient();
 
   return req.user;
