--- backend/src/common/guard/jwt-auth/index.ts
@@ -1,4 +1,0 @@
-export * from './jwt-auth.guard';
-export * from './jwtPayload';
-export * from './req-jwt-payload';
-export * from './socket-jwt-payload';
